Move string rendering function to own module
