+++ src/renderer/mod.rs
@@ -11,6 +11,7 @@ pub mod console;
 mod guidelines;
 pub mod points; // point drawing functions
                 // This imports calc_x, etc. which transforms coordinates between .glif and Skia
+pub mod string;
 
 mod glyph;
 pub mod viewport;
